+++ simulation/forces/external_force.c
@@ -1,9 +1,9 @@
 #include "external_force.h"
 
 // total force acting on the i-th body at the d-th dimention
-double fExternal(int i, int d, double x[N][D], double m[N])
+long double fExternal(int i, int d, long double x[N][D], long double m[N])
 {
-    double force = 0;
+    long double force = 0;
     for (int j = 0; j < N; j++)
     {
         force += fInternal(i, j, d, x, m);
