+++ simulation/runge_kutta/next_copy.h
@@ -3,6 +3,6 @@
 #ifndef NEXT_COPYN
 #define NEXT_COPYN
 
-void adv_copy(double vprime[N][D], double xprime[N][D], double v[N][D], double x[N][D], int mode, double w[N][D][4], double k[N][D][4]);
+void adv_copy(long double vprime[N][D], long double xprime[N][D], long double v[N][D], long double x[N][D], int mode, long double w[N][D][4], long double k[N][D][4]);
 
 #endif
