+++ simulation/utilities/initialize.c
@@ -1,6 +1,6 @@
 #include "initialize.h"
 
-void initialize(double energy, double t, double x[N][D], double v[N][D], double m[N])
+void initialize(long double energy, long double t, long double x[N][D], long double v[N][D], long double m[N])
 {
     //initialize starting conditions
     t = t_0;
